pyproject.toml
[tool.poetry]
name = "gen_trade"
version = "0.1.0"
description = ""
authors = ["Josh Harrison <joshharrison13@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
ipdb = "^0.13.9"
ipython = "^8.4.0"
python-binance = "^1.0.16"
ta = "^0.10.1"
pytest = "^7.1.2"
mock = "^4.0.3"
black = "^22.3.0"
pandas-ta = "^0.3.14-beta.0"
arrow = "^1.2.2"
memory-profiler = "^0.60.0"
dill = "^0.3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
